requirements.txt
opencv-python
pytesseract
tesseract
pandas
google.generativeai
python-dotenv
Flask
easyocr
langchain_google_genai